+++ extensions/pipe-commands/changelog.md
@@ -2,8 +2,9 @@
 
 ## Improve Development Experience (2022-03-13)
 
+- remove buggy file arguments
+- fix text input from clipboard
 - Use List Details to preview commands content from the management view
 - Add JSON schema for Pipe Commands and show parsing errors in UI
 - Add a link to the README in the command templates
 - Add better error reporting when running scripts
-- Add ability to create snippets and quicklinks from the selected text